# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: School Bus Tracking API
  version: 0.7.0
  description: |
    API for the School Bus Tracking admin web and the Android Driver and Parent apps.

    ## Conventions
    - Base path for business endpoints: `/api/v1`. JSON fields are `snake_case`.
    - Timestamps are RFC 3339 UTC (`2026-10-07T02:15:00Z`); calendar dates are `YYYY-MM-DD`.
    - Mobile numbers are accepted as `9876543210`, `+91 98765 43210`, etc. and always returned as `+919876543210`.
    - Success responses wrap the payload: `{"data": ...}`. Lists add
      `"meta": {"page", "page_size", "total"}` and accept `?page=1&page_size=20` (max 100).
    - Errors: `{"error": {"code", "message", "fields"}}`. Branch on `code`, show `message`.
    - Every response carries an `X-Request-ID` header; include it when reporting a bug.

    ## Authentication
    - **Driver and Parent apps:** `POST /auth/otp/send` then `POST /auth/otp/verify` with `app` = `driver` or `parent`.
      On the dev environment every code is `123456` (no SMS is sent).
    - **Admin web:** `POST /auth/login` (email + password). If the response has `mfa_required: true`,
      call `POST /auth/login/2fa` with the `mfa_token` and the authenticator code.
    - Send `Authorization: Bearer <access_token>` on every other call.
    - Access tokens last `expires_in` seconds (15 min). On `401`, call `POST /auth/refresh` once with the
...
      latest one. If refresh fails, send the user to login.
...

...

    ### Driver app: sending GPS
...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/schools/{schoolID}/settings/maps:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
    get:
      tags: [Settings]
      summary: The school's Google Maps keys (staff)
      description: The browser key is returned; the server key never is (only whether one is set and its last 4 characters).
      operationId: getMapsSettings
      responses:
        "200":
          description: Settings.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/MapsSettings" } } }
    put:
      tags: [Settings]
      summary: Set the school's Google Maps keys (Super Admin, School Admin)
      description: "`server_key` omitted or null keeps the stored key; an empty string removes it. Stored encrypted."
      operationId: updateMapsSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                browser_key: { type: string, description: Empty to use OpenStreetMap. }
                server_key: { type: [string, "null"] }
      responses:
        "200":
          description: Saved.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/MapsSettings" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SchoolID: { name: schoolID, in: path, required: true, schema: { type: string, format: uuid } }
...
  schemas:
...
    MapsSettings:
      type: object
      properties:
        browser_key: { type: string }
        server_key_set: { type: boolean }
        server_key_hint: { type: string, example: "…Wxyz" }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            fields:
              type: object
              additionalProperties: { type: string }
...
  responses:
...
    ValidationFailed:
      description: Invalid input.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: validation_failed, message: One or more fields are invalid., fields: { mobile: must be a valid 10-digit mobile number } } }
...